First Public Firmware - Version 1.0513
### Source Code Version 1.0513
This is the first public firmware version. It has worked with the ESP32 as is. The correct versions of the companion libraries need to be installed in your Arduino environment and you will need ESP32 support for it to compile using the Arduino IDE. All source code is contained in a single file for now. Your mileage may vary - use at your own risk and don't come crying if your cat explodes or anything else unsavory happens.